Fix phpstan issues on ajax directory

+++ ajax/taskjobdeletetype.php
@@ -42,5 +42,5 @@ $pfTaskjob = new PluginGlpiinventoryTaskjob();
 $pfTaskjob->deleteitemtodefatc(
     filter_input(INPUT_POST, "type"),
     filter_input(INPUT_POST, filter_input(INPUT_POST, "type") . 'item'),
-    filter_input(INPUT_POST, "taskjobs_id")
+    (int)filter_input(INPUT_POST, "taskjobs_id")
 );
